backend/requirements.txt
flask
flask-cors
keras
numpy
pillow
opencv-python-headless
tensorflow
